Make announcers Obj-C friendly.
I have splitted `CKComponentAnnouncerBase` into several files in order to make possible to use announcers in pure Obj-C (as opposed to Obj-C++).
Previouly `CKComponentAnnouncerBase.h` contained C++ code, which is no longer true.
